pyproject.toml
[project]
name = "geap-tour"
version = "0.1.0"
description = "GEAP Workshop: Multi-agent systems with ADK, MCP, evaluation, and observability on Google Cloud"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-aiplatform[adk,agent_engines]>=1.88.0",
    "google-genai>=1.14.0",
    "fastmcp>=2.0.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
    "google-cloud-monitoring>=2.0.0",
    "google-cloud-bigquery>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
